+++ cmd/hidden_lake/messenger/internal/config/config.go
@@ -20,11 +20,12 @@ var (
 type SConfig struct {
 	settings.SConfigSettings
 
-	FLogging    []string  `json:"logging,omitempty"`
-	FLanguage   string    `json:"language,omitempty"`
-	FAddress    *SAddress `json:"address"`
-	FConnection string    `json:"connection"`
-	FStorageKey string    `json:"storage_key,omitempty"`
+	FLogging           []string  `json:"logging,omitempty"`
+	FLanguage          string    `json:"language,omitempty"`
+	FAddress           *SAddress `json:"address"`
+	FConnection        string    `json:"connection"`
+	FStorageKey        string    `json:"storage_key,omitempty"`
+	FBackupConnections []string  `json:"backup_connections,omitempty"`
 
 	fFilepath string
 	fMutex    sync.Mutex
@@ -135,6 +136,13 @@ func (p *SConfig) GetLanguage() utils.ILanguage {
 	defer p.fMutex.Unlock()
 
 	return p.fLanguage
+}
+
+func (p *SConfig) GetBackupConnections() []string {
+	p.fMutex.Lock()
+	defer p.fMutex.Unlock()
+
+	return p.FBackupConnections
 }
 
 func (p *SConfig) GetAddress() IAddress {