+++ cmd/hidden_lake/messenger/internal/config/types.go
@@ -12,6 +12,7 @@ type IWrapper interface {
 }
 
 type IEditor interface {
+	UpdateBackupConnections([]string) error
 	UpdateLanguage(utils.ILanguage) error
 }
 
@@ -22,6 +23,7 @@ type IConfig interface {
 	GetLogging() logger.ILogging
 	GetConnection() string
 	GetStorageKey() string
+	GetBackupConnections() []string
 	GetLanguage() utils.ILanguage
 }