+++ cmd/hidden_lake/messenger/pkg/app/state/updater.go
@@ -38,7 +38,7 @@ func copyStorageState(pStorageValue *SStorageState) *SStorageState {
 	copyStorageValue := &SStorageState{
 		FPrivKey:     pStorageValue.FPrivKey,
 		FFriends:     make(map[string]string, len(pStorageValue.FFriends)),
-		FConnections: make([]*SConnection, 0, len(pStorageValue.FConnections)),
+		FConnections: make([]string, 0, len(pStorageValue.FConnections)),
 	}
 
 	for aliasName, pubKey := range pStorageValue.FFriends {