+++ cmd/hidden_lake/service/internal/handler/config_connects.go
@@ -8,6 +8,7 @@ import (
 	"github.com/number571/go-peer/cmd/hidden_lake/service/internal/config"
 	"github.com/number571/go-peer/internal/api"
 	"github.com/number571/go-peer/pkg/network/anonymity"
+	"github.com/number571/go-peer/pkg/stringtools"
 )
 
 func HandleConfigConnectsAPI(pWrapper config.IWrapper, pNode anonymity.INode) http.HandlerFunc {
@@ -32,33 +33,27 @@ func HandleConfigConnectsAPI(pWrapper config.IWrapper, pNode anonymity.INode) ht
 		connect := string(connectBytes)
 		switch pR.Method {
 		case http.MethodPost:
-			connects := append(pWrapper.GetConfig().GetConnections(), connect)
+			connects := stringtools.UniqAppendToSlice(
+				pWrapper.GetConfig().GetConnections(),
+				connect,
+			)
 			if err := pWrapper.GetEditor().UpdateConnections(connects); err != nil {
 				api.Response(pW, http.StatusInternalServerError, "failed: update connections")
 				return
 			}
-			pNode.GetNetworkNode().AddConnect(connect)
+			_ = pNode.GetNetworkNode().AddConnect(connect) // connection may be refused (closed)
 			api.Response(pW, http.StatusOK, "success: update connections")
 		case http.MethodDelete:
-			connects := deleteConnect(pWrapper.GetConfig(), connect)
+			connects := stringtools.DeleteFromSlice(pWrapper.GetConfig().GetConnections(), connect)
 			if err := pWrapper.GetEditor().UpdateConnections(connects); err != nil {
 				api.Response(pW, http.StatusInternalServerError, "failed: delete connection")
 				return
 			}
-			pNode.GetNetworkNode().DelConnect(connect)
+			if err := pNode.GetNetworkNode().DelConnect(connect); err != nil {
+				api.Response(pW, http.StatusInternalServerError, "failed: del connection")
+				return
+			}
 			api.Response(pW, http.StatusOK, "success: delete connection")
 		}
 	}
-}
-
-func deleteConnect(pCfg config.IConfig, pConnect string) []string {
-	connects := pCfg.GetConnections()
-	result := make([]string, 0, len(connects))
-	for _, conn := range connects {
-		if conn == pConnect {
-			continue
-		}
-		result = append(result, conn)
-	}
-	return result
 }