+++ pkg/stringtools/stringtools_test.go
@@ -1,0 +1,5 @@
+package stringtools
+
+import "testing"
+
+func TestNothing(t *testing.T) {}